requirements.txt
streamlit
numpy==1.26.4
pandas
matplotlib
reportlab
tensorflow-cpu==2.15.0